requirements.txt
streamlit
torch
transformers
pdfminer.six